chore : 초기 구조 생성
초기 구조 생성
